+++ ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for dwproxy
 
+## 0.1.1.0  -- 2018-12-15
+
+* Item search in shops.
+* Minor speedwalking adjustments.
+
+
 ## 0.1.0.0  -- 2018-09-30
 
 * First release: Rudimentary GMCP support, basic speedwalking between